+++ packages/viewer/src/index.ts
@@ -17,5 +17,7 @@ export {
 } from './lib/materials'
 export { mergedOutline } from './lib/merged-outline-node'
 export { default as useViewer } from './store/use-viewer'
+export { useItemLightPool } from './store/use-item-light-pool'
+export type { LightRegistration } from './store/use-item-light-pool'
 export { InteractiveSystem } from './systems/interactive/interactive-system'
 export { snapLevelsToTruePositions } from './systems/level/level-utils'